Move init_dist_profile from init::v0 to init::profile.
